Improve route parameters handling and update version
The route parameters handling has been improved to avoid setting a default requirement when there is a requirement defined in the route, in Router/Routes/Route.php file. Also, if a parameter requirement is already defined in the $requirements array, it won't be overwritten. Meanwhile, the project version has been increased to 1.6.2 in composer.json.