Updated LazyLoadRepo constants' data types and added setter method
In the LazyLoadRepo class, the data types of the constant variables have been updated from simple constants to specifically typed constants. In addition, a new function setDafaultRepoType was added to enable explicit setting of the default repository type. Also, the project version in the composer.json file was incremented to 1.6.8.