Add method to recursively merge objects in Arr.php
A new static function, mergeObjectsRecursiveDistinct, has been added to the Arr.php file, allowing for the recursive merging of two stdClass objects while removing duplicate values. This prevents repeated data and ensures the integrity of the merged object. In addition, the project version in composer.json has been updated to 1.7.1.