Add translation settings snapshot to Translatable class
This commit introduces a snapshot feature to the Translatable class that allows for the easy restoration of translation settings. The code provides methods to both set and restore a snapshot of the current settings. Additionally, minor enhancements have been made to the Arr library and the project version has been updated in composer.json.