Update version and improve ChangeHistoryTrait logic
The version number in composer.json has been updated. Additionally, the logic in the ChangeHistoryTrait has been enhanced to not only check if the class is an instance of Entity, but also if it's an instance of ValueObject. This change provides more flexibility when dealing with different types of instances.